requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-filter==25.1
djangorestframework==3.15.2
idna==3.10
Markdown==3.7
PaycomUz==3.1
payme-pkg==3.0.22
pillow==11.1.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
